Integrate promotion into sg
This adds some significant complexity, so I thought a lot if we really want this im sg. But anyone who wants to use this library will want promotions, so it's probably better to handle it here. This is also much easier for us in lila, we won't have to handle callbacks inside callbacks, and reset the whole sg config while promotion is cancelled.